#297 move resources to assets
