+++ app/src/assets/resources/locations/index.js
