+++ app/src/assets/resources/locations/latloncitysize_enrichment.py
