+++ app/src/assets/resources/tags/index.js
