+++ app/src/services/LocationService.ts
@@ -1,5 +1,5 @@
 import Location from "@/models/location";
-import locationArray from "@/resources/locations/index";
+import locationArray from "@/assets/resources/locations/index";
 import CountryService from "@/services/CountryService";
 
 class LocationService {
